backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
motor==3.7.1
pymongo==4.13.2
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0,<4.1.0
boto3==1.39.14
pydantic==2.11.7
pydantic-settings==2.10.1
pycryptodome==3.23.0
email-validator==2.2.0